moved add todo to its own component
